Created new class to hold both sides of the read
This class will be holding the methods I've been working on that edit the base reads and their quality scores. Currently has a constructor that you need to enter (baseRead1, qualRead1, baseRead2, qualRead2). After creating the pair object you can operate on it as needed with the tStrip (reduce qual of leading T and trailing 2 T's to 2), oCheck, which returns the largest section of overlap found, if it is at least 11, and the final qualPass, currently just printing a pass/fail for each read. These will all be brought together later in development, but I thought this would be a good way to organize the string manipulations and tests.