Added lAdap and rAdap
Because we couldn't see where the result files were going, we tried adding these variables to be sent to adna.cpp to be able to have some sort of print statement of what we've totally removed. If this doesn't work we will need to roll this back to the version you had working.